PMG_BS_Loader added to read a bs obj directly Pymatgen_loader name changed in PMG_Vasprun_Loader
